+++ context.cc
@@ -1,3 +1,5 @@
+#include "context.h"
+#include "deps/include/v8-external.h"
 #include "deps/include/v8-template.h"
 
 #include "context-macros.h"
@@ -27,11 +29,13 @@ ContextPtr NewContext(IsolatePtr iso,
   // side to lookup the context in the context registry. We use slot 1 as slot 0
   // has special meaning for the Chrome debugger.
   Local<Context> local_ctx = Context::New(iso, nullptr, global_template);
-  local_ctx->SetEmbedderData(1, Integer::New(iso, ref));
 
   m_ctx* ctx = new m_ctx;
   ctx->ptr.Reset(iso, local_ctx);
   ctx->iso = iso;
+  local_ctx->SetEmbedderData(ContextDataIndex::REF, Integer::New(iso, ref));
+  local_ctx->SetEmbedderData(ContextDataIndex::WRAPPER_PTR,
+                             External::New(iso, ctx));
   return ctx;
 }
 
@@ -54,6 +58,15 @@ void ContextFree(ContextPtr ctx) {
   }
 
   delete ctx;
+}
+
+m_value* track_value(m_ctx* ctx, Local<Value> value) {
+  m_value* val = new m_value;
+  val->id = 0;
+  val->iso = ctx->iso;
+  val->ctx = ctx;
+  val->ptr = Global<Value>(ctx->iso, value);
+  return tracked_value(ctx, val);
 }
 
 m_value* tracked_value(m_ctx* ctx, m_value* val) {
@@ -118,4 +131,10 @@ RtnValue RunScript(ContextPtr ctx, const char* source, const char* origin) {
 
   rtn.value = tracked_value(ctx, val);
   return rtn;
+}
+
+m_ctx* m_ctx::FromV8Context(v8::Local<v8::Context> ctx) {
+  return (m_ctx*)ctx->GetEmbedderData(ContextDataIndex::WRAPPER_PTR)
+      .As<External>()
+      ->Value();
 }